MAINT: replace /users/me route with toplevel /me route
